+++ scripts/sync_hal_source.sh
@@ -25,6 +25,7 @@ FILES=(
   "Hal Universal/QueryExpansion.swift"
   "Hal Universal/PromptDetailView.swift"
   "Hal Universal/SelfKnowledgeEngine.swift"
+  "Hal Universal/TraitCrystallizer.swift"
   "Hal Universal/Hal.swift"
 )